refactor(architecture): move files
Move `dictionaries.dart` and `settings.dart` to `repositories` directory, they should be repository originally.

Move all files under `lib/widget` to `/lib/ui/core`, because we should not make a confusion with Flutter's built-in concept.
