+++ lib/main.dart
@@ -3,21 +3,21 @@ import "dart:io";
 import "package:ciyue/database/app/app.dart";
 import "package:ciyue/database/app/daos.dart";
 import "package:ciyue/localization_delegates.dart";
-import "package:ciyue/pages/main/main.dart";
-import "package:ciyue/pages/main/wordbook.dart";
-import "package:ciyue/pages/manage_dictionaries/main.dart";
-import "package:ciyue/pages/manage_dictionaries/properties.dart";
-import "package:ciyue/pages/manage_dictionaries/settings_dictionary.dart";
-import "package:ciyue/pages/settings/ai_settings.dart";
-import "package:ciyue/pages/settings/appearance_settings.dart";
-import "package:ciyue/pages/settings/audio.dart";
-import "package:ciyue/pages/settings/auto_export.dart";
-import "package:ciyue/pages/settings/privacy_policy.dart";
-import "package:ciyue/pages/settings/terms_of_service.dart";
-import "package:ciyue/pages/word_display.dart";
-import "package:ciyue/services/dictionary.dart";
+import "package:ciyue/ui/pages/main/main.dart";
+import "package:ciyue/ui/pages/main/wordbook.dart";
+import "package:ciyue/ui/pages/manage_dictionaries/main.dart";
+import "package:ciyue/ui/pages/manage_dictionaries/properties.dart";
+import "package:ciyue/ui/pages/manage_dictionaries/settings_dictionary.dart";
+import "package:ciyue/ui/pages/settings/ai_settings.dart";
+import "package:ciyue/ui/pages/settings/appearance_settings.dart";
+import "package:ciyue/ui/pages/settings/audio.dart";
+import "package:ciyue/ui/pages/settings/auto_export.dart";
+import "package:ciyue/ui/pages/settings/privacy_policy.dart";
+import "package:ciyue/ui/pages/settings/terms_of_service.dart";
+import "package:ciyue/ui/pages/core/word_display.dart";
+import "package:ciyue/repositories/dictionary.dart";
 import "package:ciyue/services/platform.dart";
-import "package:ciyue/services/settings.dart";
+import "package:ciyue/repositories/settings.dart";
 import "package:ciyue/services/updater.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:ciyue/viewModels/audio.dart";
