+++ lib/repositories/dictionary.dart
@@ -2,11 +2,11 @@ import "dart:convert";
 import "dart:io";
 
 import "package:ciyue/database/app/app.dart";
-import "package:ciyue/pages/manage_dictionaries/main.dart";
+import "package:ciyue/ui/pages/manage_dictionaries/main.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:ciyue/utils.dart";
 import "package:ciyue/viewModels/audio.dart";
-import "package:ciyue/widget/loading_dialog.dart";
+import "package:ciyue/ui/pages/core/loading_dialog.dart";
 import "package:dict_reader/dict_reader.dart";
 import "package:drift/drift.dart";
 import "package:file_selector/file_selector.dart";
