+++ lib/services/updater.dart
@@ -1,7 +1,7 @@
 import "package:ciyue/main.dart";
 import "package:ciyue/models/updater/updater.dart";
-import "package:ciyue/services/settings.dart";
-import "package:ciyue/widget/update_available.dart";
+import "package:ciyue/repositories/settings.dart";
+import "package:ciyue/ui/pages/core/update_available.dart";
 import "package:dio/dio.dart";
 import "package:flutter/material.dart";
