+++ lib/ui/pages/core/word_display.dart
@@ -3,17 +3,17 @@ import "dart:io";
 import "dart:ui" as ui;
 
 import "package:ciyue/main.dart";
-import "package:ciyue/pages/main/wordbook.dart";
+import "package:ciyue/ui/pages/main/wordbook.dart";
 import "package:ciyue/services/audio.dart";
 import "package:ciyue/services/backup.dart";
-import "package:ciyue/services/dictionary.dart";
-import "package:ciyue/services/settings.dart";
+import "package:ciyue/repositories/dictionary.dart";
+import "package:ciyue/repositories/settings.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:ciyue/viewModels/audio.dart";
 import "package:ciyue/viewModels/home.dart";
-import "package:ciyue/widget/ai_markdown.dart";
-import "package:ciyue/widget/search_bar.dart";
-import "package:ciyue/widget/tags_list.dart";
+import "package:ciyue/ui/pages/core/ai_markdown.dart";
+import "package:ciyue/ui/pages/core/search_bar.dart";
+import "package:ciyue/ui/pages/core/tags_list.dart";
 import "package:flutter/material.dart";
 import "package:flutter/services.dart";
 import "package:flutter_inappwebview/flutter_inappwebview.dart";
