+++ lib/ui/pages/main/ai_translate_page.dart
@@ -1,7 +1,7 @@
 import "dart:ui" as ui;
 
 import "package:ciyue/services/ai.dart";
-import "package:ciyue/services/settings.dart";
+import "package:ciyue/repositories/settings.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:flutter/material.dart";
 import "package:gpt_markdown/gpt_markdown.dart";
