+++ lib/ui/pages/main/main.dart
@@ -1,7 +1,7 @@
-import "package:ciyue/pages/main/ai_translate_page.dart";
-import "package:ciyue/pages/main/home.dart";
-import "package:ciyue/pages/main/settings.dart";
-import "package:ciyue/pages/main/wordbook.dart";
+import "package:ciyue/ui/pages/main/ai_translate_page.dart";
+import "package:ciyue/ui/pages/main/home.dart";
+import "package:ciyue/ui/pages/main/settings.dart";
+import "package:ciyue/ui/pages/main/wordbook.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:ciyue/viewModels/home.dart";
 import "package:flutter/material.dart";
