+++ lib/ui/pages/main/settings.dart
@@ -5,12 +5,12 @@ import "package:ciyue/database/app/app.dart";
 import "package:ciyue/main.dart";
 import "package:ciyue/services/backup.dart";
 import "package:ciyue/services/platform.dart";
-import "package:ciyue/services/settings.dart";
+import "package:ciyue/repositories/settings.dart";
 import "package:ciyue/services/updater.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:ciyue/viewModels/home.dart";
-import "package:ciyue/widget/alpha_text.dart";
-import "package:ciyue/widget/update_available.dart";
+import "package:ciyue/ui/pages/core/alpha_text.dart";
+import "package:ciyue/ui/pages/core/update_available.dart";
 import "package:file_selector/file_selector.dart";
 import "package:flutter/material.dart";
 import "package:flutter/services.dart";
