+++ lib/ui/pages/manage_dictionaries/properties.dart
@@ -1,4 +1,4 @@
-import "package:ciyue/services/dictionary.dart";
+import "package:ciyue/repositories/dictionary.dart";
 import "package:flutter/material.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:go_router/go_router.dart";
