+++ lib/ui/pages/settings/auto_export.dart
@@ -2,7 +2,7 @@ import "dart:io";
 
 import "package:ciyue/main.dart";
 import "package:ciyue/services/platform.dart";
-import "package:ciyue/services/settings.dart";
+import "package:ciyue/repositories/settings.dart";
 import "package:ciyue/src/generated/i18n/app_localizations.dart";
 import "package:file_selector/file_selector.dart";
 import "package:flutter/material.dart";
